+++ TeachingRecordSystem/src/TeachingRecordSystem.Core/DataStore/Postgres/Migrations/20250425142820_PersonQtsDateAndEytsDate.cs
@@ -1,0 +1,39 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace TeachingRecordSystem.Core.DataStore.Postgres.Migrations
+{
+    /// <inheritdoc />
+    public partial class PersonQtsDateAndEytsDate : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<DateOnly>(
+                name: "eyts_date",
+                table: "persons",
+                type: "date",
+                nullable: true);
+
+            migrationBuilder.AddColumn<DateOnly>(
+                name: "qts_date",
+                table: "persons",
+                type: "date",
+                nullable: true);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "eyts_date",
+                table: "persons");
+
+            migrationBuilder.DropColumn(
+                name: "qts_date",
+                table: "persons");
+        }
+    }
+}