+++ TeachingRecordSystem/src/TeachingRecordSystem.SupportUi/Infrastructure/Security/Requirements/HangFireRequirement.cs
@@ -1,7 +1,0 @@
-using Microsoft.AspNetCore.Authorization;
-
-namespace TeachingRecordSystem.SupportUi.Infrastructure.Security.Requirements;
-
-public class HangFireRequirement : IAuthorizationRequirement
-{
-}