+++ CHANGELOG.md
@@ -3,14 +3,18 @@
 ## vNext
 
 The following new endpoints have been added:
-- `PUT /v3/persons/<trn>` - to set a person's PII.
 - `PUT /v3/persons/<trn>/welsh-induction` - to set a person's induction for teachers in Wales.
-- `PUT /v3/persons/<trn>/professional-statuses/<id>` - to set a professional status.
+
+## 20250425
+
+The following new endpoints have been added:
+- `PUT /v3/persons/<trn>` - to set a person's PII.
+- `PUT /v3/persons/<trn>/professional-statuses/<reference>` - to set a professional status.
 
 The `GET /v3/trn-requests` and `POST /v3/trn-requests` endpoints return the following additional properties:
 - `potentialDuplicate`;
 - `accessYourTeachingQualificationsLink`.
-`accessYourTeachingQualificationsLink` will only be populated for `Completed` requests.
+  `accessYourTeachingQualificationsLink` will only be populated for `Completed` requests.
 
 ## 20250327